turtle_inc/requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
dnspython==2.4.2
psycopg2_binary==2.9.8
pymongo==4.5.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0